Updated and fixed yet another test case: Rans/Pipe_Re_Tau_590 -> Rans/Pipe_Re_Tau_550
Previously this test did not even work and
referece data was obtained for various different Re

Now Re = U_b*D/nu = 19000 or Re_tau = 550
Reference data is supplied in Xmgrace folder and
was recieved by Fisher on nek5000 on the same Re
All necessary info is provided in
Xmgrace/Re_tau_550_ref_data.dat header

Also, corresponding python_matplotlib_script.py is updated
It is a bit different from the one for Channel flow since
Rustam was interested on the Uz(r) as well.

It seems that majority of RANS models are doing fine
in the viscous layer, but not in the center of a pipe.
The only exception - rsm_hanjalic_jakirlic model.