A bit more elaborate opening message.
Checks if the file existed, and if no it says it is
creating the file, if it did, it says it is appending.

(This was to decrease the excessive messages from the
Grid_Mod_Exchange_Real when writing info on redunancy.

On branch membrane

modified:   Shared/File_Mod/Append_File_For_Writing.f90