IMPORTANT: A resolution of a very subtle bug!
Since the introduction of face-based body forces, a very slight
imbalance in pressure source can occur globally (over all processors)
and we are talking about the values of the order 1e-18 here.  This
happens because of the round-off errors associated with real arithmetics
performed in CPUs, meaning that: A + B = B + A + epsilon.  If forces
are computed over faces, they are slightly different in different
processors due to these round-off errors, although values in cells
surrounding them are the same after exchanging the buffers.  In order
to fix it, a balancing procedure of pressure sources over all CPUs
has been introduced.

On branch membrane

modified:   Compute_Pressure.f90