IMPORTANT: Rhie_And_Chow handles surface tension!
After a lot of reading on this issue, and reading about it,
I focused on this equation (I think it is Laplace's equation):

sigma kappa dc/dx = dp/dx  [N/m^3]

and realized it can be satisfied in case of quiscent flow only
if c and p are numerically treated in the same way.  Hence, not
only the same method for gradients as often cited in literature,
but the whole story: where is defined, so how is interpolated at
faces in addition to gradient methods.

In order to enhance that coupling, I decided to merge together
Vof % Pressure_Correction with Rhie_And_Chow.  That will give the
strongest possible link.

With this commit, that is handled and Vof % Pressure_Correction
is not even compiled any more :-)

On branch membrane

modified:   Compute_Pressure.f90
modified:   Rhie_And_Chow.f90
modified:   Vof_Mod.f90