IMPORTANT: MatSetValue must use global numbers!
This is what I was suspecting, but now I know for
sure.  When seting values for a matrix in parallel,
rows and columns must be set with global indicese!

(That must be the way how PETSc eventually decides
on domain decomposition - if it is independent from
the rest of the code - and works out communication
between processors.)

Anyhow, with this fix, it managed to set values for
a lid driven cavity domain decomposed in two pro-
cessors only.  For the jet, parallel version still
doesn't work.

On branch bojan_uses_mpi_mod_integers_single

modified:   Petsc.True