Added support for Intel compiler.
A couple of fixes were needed for that, but the most
annoying was that Intel suite of compiler tools comes
with its own Python installation, which does not work
the same as the default system one.  So ... well ...
plotting is currently disabled for tests with Intel
compiler.

(This is not such a great deal because I rarely look
at the plots anyhoo.)

On branch adjust_first_dim

modified:   Tests/test_build.sh