Checks Front % b_node_1(:), Front % b_node_2(:)
These two fields store global node numbers whichi were used
for vertex creation.  At the moment, this is just for an
addiional check with Assert, but in the future it could and
should be used instead of coordinate comparisons.  This is
a first step in that direction.

On branch boil_bojan

modified:   Compress_Front_Vertices.f90