Further improvements of speed of conversion
This was achieved by replacing "File % Read_Line" with
as many simple and plain Fortran's "read" statements.

The improvement was staggering.  As an illustration.  For a typical
grid with 1M cells, the reading of GMSH files was reduced from 1 minute
and 20 seconds to only 18 seconds. The former timing was measured for
check in: 95aad328063b80abb39394a9dad933e5104e3b3e, not just for the
previous check in.

It would be useful to make similar changes to Load_Fluent.f90

On branch bojan_fixed_cell_ranges

modified:   Convert/Convert_Mod/Load_Gmsh.f90