IMPORTANT: Spud up native linear solvers by 14%
This speed-up was measured on a system with half a million cells.
It took me long enough.  So, I removed the preconditioning matrix
"D" and replaced it with two vectors: "d" and "d_inv" holding just
the diagonal of the preconditioning matrix.

The rationale was the following: The preconditioning matrix was
originally introduced as a Matrix_Type in case I want to expand
the preconditioning with more levels of fill-in.  But, since the
introduction of PETSc libraries, that hardly makes sense, it is
highly likely that PETSc comes with such preconditioner and if
it doesn't. there must be a good reason for that.

All in all: addressing of a vector "d" should be faster than in-
direct addressing of the matrix "D", plus the "d_inv" saves from
a lot of divisions.

On branch bojan_fixed_cell_ranges

modified:   Native_Mod.f90
modified:   Native_Mod/Bicg.f90
modified:   Native_Mod/Cg.f90
modified:   Native_Mod/Create_Native.f90
modified:   Native_Mod/Prec_Form.f90
modified:   Native_Mod/Prec_Solve.f90