An error has be reduced to warning.
In cases of simulaitons with multiple domains, in particulart when
mutliple domains are meshed concurrently in GMSH to ensure comformity
but some are deleted just before saving the files, the error picked
here can occur.  However, in that case it is really not an error but
rather some faces are hanging blank on purpose.

On branch bojan_more_robust_communicator

modified:   Convert_Mod/Find_Parents.f90