Resolved compiler warning for Tokenizer again
OK, so it was not the size of the Tokenizer about which compiler
was complaining, but the size of the Tokenizer when it is inisde
another procedure, this time it was Frameless in Message_Mod.

To resolve the warning, I ditched the local Tokenizer from Frameless
and use global (singleton) one called "Line".  Also, the definition
of the line had to be moved, from File_Mod to Tokenizer_Mod, where it
should have naturally been all along.

On branch bojan_more_robust_communicator

modified:   Shared/File_Mod.f90
modified:   Shared/Message_Mod/Frameless.f90
modified:   Shared/Tokenizer_Mod.f90