IMPORTANT: changed the orientation of cells' faces
when extracting cells from grid.  The way it was before, the Front
elements were pointing from VOF=1 towards VOF=0, which is wrong.

Read also the comments close to the part of the code which changed.
There is no doubt that the whole thing is still a bit nebulous, not
surprising when one uses third party libraries, they are always
shrouded in a bit of mystery.

On branch bojan_more_robust_communicator

modified:   Shared/Polyhedron_Mod/Extract_From_Grid.f90