Added the following note to makefile for Process:
At the very end of the makefile, environment variables for PETSc
libraries are included.  If you have a compiler mismatch there,
that is if you PETSc was compiled with different compiler from
the one you are using to compile Process, errors will occur.

On branch bojan_iter_module

modified:   Sources/Process/makefile