PART OF THE ATTEMPT TO IMPROVE PETSC PERFORMANCE
On branch bojan_tuning_petsc_again

One of the most important changes I did in this endeavour
was to introuce Petsc_Type inside a variable, so that each
variable can hold its own copy of PETSc matrices, vectors
and solvers.  (By doing so, I was hoping to improve the
performance of the code by avoiding copying the matrices and
updating the preconditioner, but it lead to very modest
improvements.)

Changes done:

1. modified:   Var_Mod.f90
   renamed:    Var_Mod/Allocate_New_Only.f90 -> Var_Mod/Create_New_Only.f90
   renamed:    Var_Mod/Allocate_Solution.f90 -> Var_Mod/Create_Solution.f90
   new file:   Var_Mod/Destroy_New_Only.f90
   new file:   Var_Mod/Destroy_Solution.f90

I introduced Petsc_Type inside the variable, but also added
a few member functions (to destroy the object, and renamed
some.

2. modified:   Solver_Mod.f90
   modified:   Solver_Mod/Create_Solver.f90
   modified:   Solver_Mod/End.f90
   modified:   Solver_Mod/Remove_Singular.f90
   modified:   Solver_Mod/Run.f90
   modified:   Solver_Mod/Set_Singular.f90
   modified:   Petsc_Mod.f90
   modified:   Petsc_Mod/Fake/Create_Petsc.f90
   new file:   Petsc_Mod/Fake/Destroy_Petsc.f90
   modified:   Petsc_Mod/Fake/Solve_Petsc.f90
   modified:   Petsc_Mod/True/Create_Petsc.f90
   new file:   Petsc_Mod/True/Destroy_Petsc.f90
   modified:   Petsc_Mod/True/Solve_Petsc.f90

Since the Petsc_Type changed its position, it reflected the way Solver_Mod
is written.  The definition of Petsc_Type is gone and syntax to call solvers
changed too.  Changes to Petsc_Mod were necessary too.

3. modified:   Field_Mod.f90
   renamed:    Field_Mod/Core/Allocate_Field.f90 -> Field_Mod/Core/Create_Field.f90
   modified:   Field_Mod/Utilities/Compute_Wall_Distance.f90
   modified:   Field_Mod/Utilities/Potential_Initialization.f90

Field_Mod, since it holds a Var_Type, was impacted by the changes in Var_Mod

4. modified:   Process_Mod/Compute_Momentum.f90
   modified:   Process_Mod/Compute_Pressure.f90
   modified:   Process_Mod/Compute_Scalar.f90
   modified:   Process_Mod/Compute_Energy.f90
   modified:   Turb_Mod/Compute_Variable.f90
   modified:   Numerics_Mod/Advection_Term.f90

In these files, only the position of the logical blend_matrix was moved from
the Field_Mod to the Var_mod.

5. modified:   Main_Pro.f90
   modified:   Swarm_Mod.f90
   renamed:    Swarm_Mod/Allocate_Swarm.f90 -> Swarm_Mod/Create_Swarm.f90
   modified:   Turb_Mod.f90
   modified:   Turb_Mod/Compute_Stress.f90
   modified:   Turb_Mod/Compute_Variable.f90
   renamed:    Turb_Mod/Allocate_Turb.f90 -> Turb_Mod/Create_Turb.f90
   modified:   Vof_Mod.f90
   renamed:    Vof_Mod/Core/Allocate_Vof.f90 -> Vof_Mod/Core/Create_Vof.f90

The changes following the change in Var_Mod, but also, for the sake of unifor-
mity, I renamed all Allocate_ functions to Create_.

6. modified:   Native_Mod.f90
   modified:   Native_Mod/Create_Native.f90

Here I only added an assertion.

7. modified:   C_Petsc_Interface.c

Added a lot of new interface functions to PETSc

8. modified:   Matrix_Mod.f90
   modified:   Matrix_Mod/Create_Matrix.f90

Moved array "glo" from Petsc_Mod to Matrix_mod.

9. modified:   Read_Controls_Mod/Numerical_Schemes.f90
   modified:   Read_Controls_Mod/Petsc_Solvers.f90
   modified:   Laminar/Backstep/Orthogonal/Solvers_Scaling/control
   modified:   Les/Matrix_Of_Cubes/control
   modified:   Manual/Matrix_Of_Cubes/Comparison/control
   modified:   Manual/Matrix_Of_Cubes/control
   modified:   Vof/Rising_Bubble/control
   modified:   Documentation/Syntax/.vim/syntax/control.vim