Neat additions to the ford_run.sh script.
Now it makes explicit exclusion of the sources you do not want
to analyze, and also has argument cheching.  This is hardly the
last version of the script, but now it is a wee-bit smarter than
it was before.

On branch big_merge_in_november_2023_attempt_01

modified:   Documentation/Ford/ford_run.sh