On the path to MPI parallel verion: Communication
This is not a fix, but a significant improvement in terms
of computational efficiency.  When determining cell patterns
for communication, we no longer browse through all the cells,
but only to cells close to buffers - which are the cells
which share at least one node with buffer cells (that't how
it is in Divide).

On branch bojan_gpu_development

modified:   Sources/Shared/Grid_Mod/Parallel/Form_Cells_Comm.f90