Added test for volume balance of rotating field
It uses the same algorithm as Insert_Volume_Source_For_Pressure.f90
and Correct_Velocity.f90, but this is here to test the MPI version.
It is very small and not intented to be used/tested on GPUs.

On branch bojan_gpu_development

modified:   Main.f90
new file:   Test_008.f90
new file:   convert_008.scr